% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{divmultBreaks}
\alias{divmultBreaks}
\title{Compute breaks for ratio scale}
\usage{
divmultBreaks(v, n = 6, nmin = 3, anchor = TRUE)
}
\arguments{
\item{v}{Numeric vector}

\item{n}{Scalar, target number of breaks}

\item{nmin}{Scalar, forced minimum number of breaks}

\item{anchor}{Logical, include origin (1 on the ratio scale)}
}
\value{
Vector of values to generate axis breaks
}
\description{
Compute breaks for ratio scale
}
\examples{
divmultBreaks(c(11))
divmultBreaks(c(0.04))
divmultBreaks(c(0.04, 11))
divmultBreaks(c(0.02, 2))
divmultBreaks(c(0.8, 20))
}
