% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-doCall2.R
\name{doCall2}
\alias{doCall2}
\title{Modified do.call for passing optional arguments to child functions}
\usage{
doCall2(what, args, accepted = methods::formalArgs(what))
}
\arguments{
\item{what}{Character, name of function}

\item{args}{List of unquoted arguments}

\item{accepted}{Character, quoted names of arguments to be passed to \code{what}}
}
\value{
Output of what
}
\description{
Modified do.call for passing optional arguments to child functions
}
\details{
source code and example from stack overflow conversation with G.
Grothendieck here: https://stackoverflow.com/a/71801214/8400969
}
\examples{

dat <- c(NA, 1:5)
meanLog <- function(x, ...){
  y <- doCall2("log", list(x, ...), "base")
  z <- doCall2("mean.default", list(y, ...))
  return(z)
  }

meanLog(dat, na.rm = TRUE, base = 2)
}
