% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exch}
\alias{exch}
\title{USD-CAD exchange rate in first two COVID pandemic years}
\format{
Data frame with 522 observations of 3 variables
\itemize{
\item \strong{date} Date, "YYYY-MM-DD"
\item \strong{CADtoUSD} Numeric, CAD per USD
\item \strong{USDtoCAD} Numeric, USD per CAD
}
}
\source{
Data from the Federal Reserve Economic Data (FRED), accessed with the
R package \strong{alfred}.
}
\usage{
exch
}
\description{
A dataset with weekday average exchange rate between Canadian and US Dollars
between 1 April 2020 and 31 March 2022.
}
\keyword{datasets}
