% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_centiNel}
\alias{label_centiNel}
\title{100x Natural log (centinel) transformation of breaks}
\usage{
label_centiNel()
}
\value{
Function used as argument to \code{labels} in \verb{scale_*_*}
}
\description{
100x Natural log (centinel) transformation of breaks
}
\seealso{
Other {labels}: 
\code{\link{label_divMult}()},
\code{\link{label_nel}()},
\code{\link{label_percDiff}()},
\code{\link{label_propDiff}()}
}
\concept{{labels}}
