% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_percDiff}
\alias{label_percDiff}
\title{Scale breakpoints based on percentage difference from reference}
\usage{
label_percDiff(logscale = FALSE, base = 10)
}
\arguments{
\item{logscale}{Logical, are breaks provided on the log scale (default is
\code{FALSE})?}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}
}
\value{
Function used as argument to \code{labels} in \verb{scale_*_*}
}
\description{
Scale breakpoints based on percentage difference from reference
}
\seealso{
Other {labels}: 
\code{\link{label_centiNel}()},
\code{\link{label_divMult}()},
\code{\link{label_nel}()},
\code{\link{label_propDiff}()}
}
\concept{{labels}}
