% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_propChange}
\alias{label_propChange}
\title{Rescaling breaks relative to reference point}
\usage{
label_propChange(ref = 1)
}
\arguments{
\item{ref}{Scalar, reference value for proportional change}
}
\value{
Function used as argument to \code{labels} in \verb{scale_*_*}
}
\description{
Rescaling breaks relative to reference point
}
