% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{limit_breaks}
\alias{limit_breaks}
\title{Truncate log-scaled axis breaks to data range}
\usage{
limit_breaks(v, n = 5, splits = 1, base = exp(1))
}
\arguments{
\item{v}{Numeric vector, data or data range}

\item{n}{Scalar, target number of breaks}

\item{splits}{Integer, one of \code{c(1,2,3)}. How many tick marks per
"decade?"}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}
}
\value{
Vector of numeric values for axis breaks
}
\description{
Truncate log-scaled axis breaks to data range
}
\examples{
dat <- exp(seq(-2,5,0.2))
v <- log(dat) # data or data range
n <- 5
# axisTicks returns values way beyond data
grDevices::axisTicks(nint = n, log = TRUE, usr = range(v))
# limit_breaks reels this in
limit_breaks(v = v, n = n)
}
