% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{limitimil}
\alias{limitimil}
\title{Expand limits for symmetry on log scale}
\usage{
limitimil(v)
}
\arguments{
\item{v}{Numeric vector of length > 1, data range or limits}
}
\value{
Numeric vector at least as long as v, with upper and lower range
limits symmetrical around 1 on log scale.
}
\description{
Expand limits for symmetry on log scale
}
\examples{

limitimil(c(0.66, 2.1))
\dontrun{
imitimil(c (-1, 3))
}

}
\seealso{
Other {breaking}: 
\code{\link{breaks_divMult}()},
\code{\link{limit_breaks}()},
\code{\link{split_decades}()}
}
\concept{{breaking}}
