% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nel_vid}
\alias{nel_vid}
\title{NY Times COVID case data for mid-Atlantic USA in first two pandemic years}
\format{
Data frame with 4368 observations of 11 variables:
\itemize{
\item \strong{date} Date, YYYY-MM-DD
\item \strong{state} Character, name of US state or territory
\item \strong{cases} Integer, reported daily COVID cases
\item \strong{cases_avg} Numeric, rolling daily average number of cases within
"state" in previous 2 weeks
\item \strong{cases_avg_per_100k} Numeric, rolling daily average case rate per 100k
residents in previous 2 weeks
\item \strong{deaths} Integer, reported daily COVID-related deaths
\item \strong{deaths_avg} Numeric, rolling daily average number of deaths within
"state" in previous 2-weeks
\item \strong{deaths_avg_per_100k} Numeric, rolling daily average death rate per 100k
residents in previous 2 weeks
\item \strong{ref_case_rate} Numeric, rolling daily average case rate per 100k
residents in reference 2-week time period
\item \strong{nel_rate} Numeric, case rate rescaled to reference and expressed in
"nels"
\item \strong{prop_rate} Numeric, case rate rescaled to reference and expressed as a
proprtion of the reference rate
}

@source  Data from The New York Times, based on reports from state and local
health agencies. \url{https://github.com/nytimes/covid-19-data}
}
\usage{
nel_vid
}
\description{
A dataset containing COVID cases and deaths for 6 US states and territories
from 1 April 2020 to 31 March 2022.
}
\keyword{datasets}
