% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{print_operator}
\alias{print_operator}
\title{Ratio labels}
\usage{
print_operator(logscale = F, base = exp(1))
}
\arguments{
\item{logscale}{Logical, is x already on the log scale?}

\item{base}{Scalar, base of the logarithm used}

\item{x}{Numeric}
}
\value{
Expression for labeling ggplots based on breaks
}
\description{
Ratio labels
}
\examples{
print_operator(((-1:3)))

}
\concept{Visualization}
