% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{propChange_trans}
\alias{propChange_trans}
\title{Natural log transformation... showing proportional change explicitly}
\usage{
propChange_trans(n = 7, split = TRUE, ref = 1, base = 10)
}
\arguments{
\item{n}{Scalar, target number of breaks}

\item{split}{logical, split decades using split_decades}

\item{ref}{Scalar, reference value for proportional change}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}
}
\description{
Natural log transformation... showing proportional change explicitly
}
\examples{
dat<-data.frame(x = 1:10, y = exp(-2:7))
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y)) +
    ggplot2::geom_point() +
    ggplot2::scale_y_continuous(
      trans = propChange_trans(ref = 25)
      , labels = label_propChange(ref = 25)
      , sec.axis = ggplot2::sec_axis(
          labels = function(x) {x}
          , trans = ~.
          , breaks = c(0.1, 0.5, 1, 5, 10, 50, 100, 500, 1000)
          , name = "original data"
        )
      ) +
    ggplot2::labs(y = "propChange scale") +
    ggplot2::geom_hline(yintercept = 25, size = 0.2)

dat \%>\% ggplot2::ggplot(ggplot2::aes(x, exp(seq(-1, 0.8, 0.2)))) +
 ggplot2::geom_point() +
 ggplot2::scale_y_continuous(
   trans = propChange_trans()
   , labels = label_propChange()
   , sec.axis = ggplot2::sec_axis(
     labels = function(x) {x}
     , trans = ~.
     , name = "original data"
   )
 ) +
 ggplot2::labs(y = "propChange scale") +
 ggplot2::geom_hline(yintercept = 1, size = 0.2)




}
