% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{rel_start}
\alias{rel_start}
\title{Rescale variables to start values by group}
\usage{
rel_start(df, col_vars, group_vars)
}
\arguments{
\item{df}{data.frame}

\item{col_vars}{variable names (quoted or unquoted) to rescale to initial
values}

\item{group_vars}{variable names (quoted or unquoted) to group \code{col_vars} by}
}
\value{
data.frame with all original data and new \code{rel_} columns for each
rescaled variable
}
\description{
Convenience wrapper of \code{dplyr::group_by} and \code{dplyr::mutate} to
scale variables, especially time-series, by series start values
}
\examples{
test_df <- data.frame(grp = rep(1:3, each = 10)
, val = rep(10:1, times = 3))


test_df \%>\%
  rel_start(col_vars = val, group_vars = grp) \%>\%
  print(n = 30)

test_df \%>\%
  rel_start(col_vars = "val", group_vars = "grp") \%>\%
  print(n = 30)
}
\concept{convenience}
