% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rockfish}
\alias{rockfish}
\title{Rockfish catch by region and vessel type}
\format{
Data frame with 234 observations of 14 variables
\itemize{
\item \strong{Estimate.Status} Character, one of "FINAL" or
\item \strong{Year}
\item \strong{Fishing.Year}
\item \strong{Region}
\item \strong{Common.Name}
\item \strong{Fishing.Mode}
\item \strong{Cumulative.Through}
\item \strong{PSE.Total.Catch..A.B1.B2.}
\item \strong{Does.Total.Catch..A.B1.B2..Meet.MRIP.Standard}
\item \strong{Is.Total.Catch..A.B1.B2..Significantly.Different.From.0}
\item \strong{Total.Catch..A.B1.B2.}
\item \strong{Total.Catch..A.B1.B2..Lower.95..Confidence.Limit}
\item \strong{Total.Catch..A.B1.B2..Upper.95..Confidence.Limit}
\item \strong{X...Contribution.of.Imputed.Data.to.Total.Catch.Rate}
}
}
\source{
Data from NOAA NMFS \url{https://www.st.nmfs.noaa.gov/SASStoredProcess/guest?_program=\%2F\%2FFoundation\%2FSTP\%2Fmrip_qry_index&qrycatch=Time+Series&subcatch=Goto+Query&qryeffort=Time+Series&qryparticipation=Select+a+Participation+Query}
}
\usage{
rockfish
}
\description{
A dataset with annual catch of striped bass ("Rockfish") in Atlantic fishing
regions between 2005 and first 3 quarters of 2023
}
\keyword{datasets}
