% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{split_decades}
\alias{split_decades}
\title{Split stingy limit_breaks into more parts per complete decade}
\usage{
split_decades(v, splits = c(0, 1, 2, 3, 4, 5, 10))
}
\arguments{
\item{v}{Vector on the unlogged scale to be examined and split}

\item{splits}{Integer, one of \code{c(1,2,3)}. How many tick marks per
"decade?"}
}
\value{
Vector with splits added
}
\description{
Split stingy limit_breaks into more parts per complete decade
}
\seealso{
Other {breaking}: 
\code{\link{breaks_divMult}()},
\code{\link{limit_breaks}()},
\code{\link{limitimil}()}
}
\concept{{breaking}}
