% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{splitDecades}
\alias{splitDecades}
\title{Split stingy limBreaks into three parts per complete decade}
\usage{
splitDecades(v)
}
\arguments{
\item{v}{vector on the unlogged scale to be examined and split}
}
\value{
vector with splits added
}
\description{
Split stingy limBreaks into three parts per complete decade
}
