% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{trans_picker}
\alias{trans_picker}
\title{Parse flags for ratio scales}
\usage{
trans_picker(tickVal, ...)
}
\arguments{
\item{tickVal}{Character, one of "divMult", "propDiff", "percDiff", "nel", or
"centiNel"}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{scale_y_continuous}} or other scale elements (e.g.,
breaks, labels, etc. )}
}
\value{
List of arguments to pass to scale_(x|y)_continuous()
}
\description{
Parse flags for ratio scales
}
