% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ushs}
\alias{ushs}
\title{US home values by region}
\format{
Data frame with 255,970 observations of 7 variables:
\itemize{
\item \strong{RegionID} Integer, numeric region identifier
\item \strong{Size_Rank} Integer, rank of region size
\item \strong{RegionName} Character, name of region
\item \strong{RegionType} Character, one of \code{c("country", "msa")}
\item \strong{StateName} Character, name of state in which region falls, if applicable
\item \strong{RecordDate} Date, time for which home values were estimated
\item \strong{TypicalHomeValue} Numeric, estimated mean value, in US dollars, of
typical homes within region and time period
}
}
\source{
Zillow Home Value Index \url{https://files.zillowstatic.com/research/public_csvs/zhvi/Metro_zhvi_uc_sfrcondo_tier_0.33_0.67_month.csv?t=1700060680}
}
\usage{
ushs
}
\description{
A dataset containing seasonally adjusted, monthly estimated typical home
values within the middle 30th percentiles (35th to 65th) for regions across
the United States
}
\keyword{datasets}
