% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{admit}
\alias{admit}
\title{Opportunity Insights data from Chetty et al. 2023 Selective Colleges
published 24 July 2023 by the NYT upshot}
\format{
Data frame with 1946 observations of 57 variables:
\itemize{
\item \strong{super_opeid} Integer, institutional ID
\item \strong{name} Character, name of college or group
\item \strong{par_income_bin} Numeric, parent income percentile bin
\item \strong{par_income_bin} Character, bin range for parental income
\item \strong{attend} Numeric, test-score-weighted attendance rate
\item \strong{stderr_attend} Numeric, standard error of attendance rate
\item \strong{attend_level} Numeric, denominator for attend rate
\item \strong{attend_sat} Numeric
\item \strong{stderr_attend_sat} Numeric
}
}
\source{
Data from Opportunity Insights \url{https://opportunityinsights.org/wp-content/uploads/2023/07/CollegeAdmissions_Data.csv}
}
\usage{
admit
}
\description{
A dataset containing college admissions rates for different socio-economic
groups
}
\keyword{datasets}
