% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{breaks_divmult}
\alias{breaks_divmult}
\title{Idiomatic breaking function for ggplot}
\usage{
breaks_divmult(n = 6, ...)
}
\arguments{
\item{n}{Integer, target number of breaks}

\item{...}{Additional arguments passed to divmultBreaks}
}
\value{
A function that takes a vector (e.g. range of data) as an argument
and returns values for breaks on a log scale, but with original scale values
}
\description{
Idiomatic breaking function for ggplot
}
\examples{
breaks_divmult()(exp(seq(0,5,0.2)))
}
