% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{divMult_trans}
\alias{divMult_trans}
\title{Natural log transformation... providing breaks on the "divMult" scale}
\usage{
divMult_trans(n = 7, base = exp(1), splits = 2, slashStar = TRUE, ...)
}
\arguments{
\item{n}{Scalar, target number of breaks}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}

\item{splits}{Integer, one of \code{c(1,2,3)}. How many tick marks per
"decade?"}

\item{slashStar}{Logical, should division and multiplication symbols be "*"
and "/" (default). Prettier symbols \eqn{\times, \div} are available when
\code{slashStar == FALSE}, but font libraries and text size may make
distinguishing \eqn{\div} from \eqn{+} difficult.}

\item{...}{Additional arguments passed to breaking function, labeller}
}
\description{
Natural log transformation... providing breaks on the "divMult" scale
}
\examples{
dat<-data.frame(x = 1:10, y = exp(-2:7))
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y)) +
  ggplot2::geom_point() +
    ggplot2::scale_y_continuous(
       trans = "divMult"
       # default breaks aren't perfect; sometimes adding more helps
       #  trans = nel_trans(n = 9)
       , labels = label_divMult()
       , sec.axis = ggplot2::sec_axis(
           labels = function(x) {x}
           , trans = ~.
           , breaks = c(0.1, 0.5, 1, 5, 10, 50, 100, 500, 1000)
           , name = "original scale"
         )
       ) +
     ggplot2::labs(y = "nel (natural log) scale") +
     ggplot2::geom_hline(yintercept = 1, linewidth = 0.2)

}
