% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{divMultBreaks}
\alias{divMultBreaks}
\title{Compute breaks for ratio scale}
\usage{
divMultBreaks(n = 6, nmin = 3, anchor = TRUE, split = FALSE)
}
\arguments{
\item{n}{Scalar, target number of breaks}

\item{nmin}{Scalar, forced minimum number of breaks}

\item{anchor}{Logical, always include origin (1 on the ratio scale)}

\item{split}{logical, split decades using splitDecades}
}
\value{
Vector of values to generate axis breaks
}
\description{
Compute breaks for ratio scale
}
\examples{
y <- exp(seq(-2,5, length.out = 10))
v <- log(y) # data or data range
n <- 5

# axisTicks takes giant steps, returns values way beyond data
grDevices::axisTicks(nint = n, log = TRUE, usr = range(v))
# divMultBreaks gives ~n breaks evenly within the data
divMultBreaks(n = n)(v = y)

# if 1 is lower limit, only positive log(breaks)
divMultBreaks()(c(1, 11))
# ditto, only negative log(breaks) if 1 is upper limit
divMultBreaks()(c(0.04, 1))

# expanding range on one side of 1 doesn't leave the other side behind
divMultBreaks()(c(0.04, 2.2))
divMultBreaks()(c(0.04, 220))
divMultBreaks()(c(0.04, 2200))

x <- 1:10
dat <- data.frame(x, y)
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y))+
     ggplot2::geom_point()+
     ggplot2::geom_hline(yintercept = 1, size = 0.2) +
     ggplot2::scale_y_continuous(
     trans = "log"
     , breaks = divMultBreaks()
     , labels = label_divMult()
     )
}
