% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exch}
\alias{exch}
\title{USD-CAD exchange rate in first two COVID pandemic years}
\format{
Data frame with 1044 observations of 4 variables
\itemize{
\item \strong{date} Date, "YYYY-MM-DD"
\item \strong{direction} Character, direction of exchange rate
\item \emph{exRate}* Numeric, exchange rate
\item \strong{exRate_scale} Numeric, proportional difference in exchnage rate from 2020-04-01
}
}
\source{
Data from the Federal Reserve Economic Data (FRED), accessed with the
R package \strong{alfred}.
}
\usage{
exch
}
\description{
A dataset with weekday average exchange rate between Canadian and US Dollars
between 1 April 2020 and 31 March 2022.
}
\keyword{datasets}
