% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_divMult}
\alias{label_divMult}
\title{Ratio labels}
\usage{
label_divMult(logscale = FALSE, base = exp(1), slashStar = TRUE)
}
\arguments{
\item{logscale}{Logical, are breaks already on the log scale?}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}

\item{slashStar}{Logical, should division and multiplication symbols be "*"
and "/" (default). Prettier symbols \eqn{\times, \div} are available when
\code{slashStar == FALSE}, but font libraries and text size may make
distinguishing \eqn{\div} from \eqn{+} difficult.}
}
\value{
Function for generating labeling expressions based on breaks
}
\description{
Ratio labels
}
\examples{
label_divMult()(c(1:4,2))

}
\seealso{
Other {labels}: 
\code{\link{label_centiNel}()},
\code{\link{label_nel}()},
\code{\link{label_percDiff}()},
\code{\link{label_propDiff}()}
}
\concept{{labels}}
