% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_divMult}
\alias{label_divMult}
\title{Ratio labels}
\usage{
label_divMult(logscale = F, base = exp(1))
}
\arguments{
\item{logscale}{Logical, are breaks already on the log scale?}

\item{base}{Scalar, base of the logarithm used}
}
\value{
Function for generating labeling expressions based on breaks
}
\description{
Ratio labels
}
\examples{
label_divMult()(c(1:4,2))

}
\concept{Visualization}
