% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{label_nel}
\alias{label_nel}
\title{Natural log (nel) transformation of breaks}
\usage{
label_nel()
}
\value{
Function used as argument to \code{labels} in \verb{scale_*_*}
}
\description{
Natural log (nel) transformation of breaks
}
\concept{labels}
