% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{nel_trans}
\alias{nel_trans}
\title{Natural log transformation... providing breaks on the "nel" scale}
\usage{
nel_trans(n = 7, use_centiNel = FALSE)
}
\arguments{
\item{n}{Integer, desired number of breaks}

\item{use_centiNel}{Logical, should units be "centiNels" (defaul is "nel")}
}
\description{
Natural log transformation... providing breaks on the "nel" scale
}
\examples{
dat<-data.frame(x = 1:10, y = exp(-2:7))
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y)) +
  ggplot2::geom_point() +
    ggplot2::scale_y_continuous(
       trans = "nel"
       # default breaks aren't perfect; sometimes adding more helps
       #  trans = nel_trans(n = 9)
       , labels = label_nel()
       , sec.axis = ggplot2::sec_axis(
           labels = function(x) {x}
           , trans = ~.
           , breaks = c(0.1, 0.5, 1, 5, 10, 50, 100, 500, 1000)
           , name = "original data"
         )
       ) +
     ggplot2::labs(y = "nel (natural log) scale") +
     ggplot2::geom_hline(yintercept = 1, size = 0.2)

}
\seealso{
\code{\link[scales]{log_breaks}}
}
