% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{propDiff_trans}
\alias{propDiff_trans}
\title{Natural log transformation... showing proportional change explicitly}
\usage{
propDiff_trans(n = 7, base = exp(1), ...)
}
\arguments{
\item{n}{Scalar, target number of breaks}

\item{base}{a positive or complex number: the base with respect to which
    logarithms are computed.  Defaults to \eqn{e}=\code{exp(1)}.}

\item{...}{additional arguments passed to \code{label_propDiff}}
}
\description{
Natural log transformation... showing proportional change explicitly
}
\examples{
dat<-data.frame(x = 1:10, y = exp(-2:7))
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y)) +
    ggplot2::geom_point() +
    ggplot2::scale_y_continuous(
      trans = propDiff_trans()
      , sec.axis = ggplot2::sec_axis(
          labels = function(x) {x}
          , trans = ~.
          , breaks = c(0.1, 0.5, 1, 5, 10, 50, 100, 500, 1000)
          , name = "original scale"
        )
      ) +
    ggplot2::labs(y = "propDiff scale") +
    ggplot2::geom_hline(yintercept = 1, size = 0.2)

dat \%>\% ggplot2::ggplot(ggplot2::aes(x, exp(seq(-1, 0.8, 0.2)))) +
 ggplot2::geom_point() +
 ggplot2::scale_y_continuous(
   trans = propDiff_trans()
   , sec.axis = ggplot2::sec_axis(
     labels = function(x) {x}
     , trans = ~.
     , breaks = c(0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2)
     , name = "original scale"
   )
 ) +
 ggplot2::labs(y = "propDiff scale") +
 ggplot2::geom_hline(yintercept = 1, size = 0.2)




}
\seealso{
\code{\link[scales]{log_breaks}}
}
