% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{scale_y_ratio}
\alias{scale_y_ratio}
\title{Y-axis wrapper for scale_ratio}
\usage{
scale_y_ratio(tickVal = "divMult", splits = 2, ...)
}
\arguments{
\item{tickVal}{Character, one of "divMult", "propDiff", "percDiff", "nel", or
"centiNel"}

\item{splits}{Integer, one of \code{c(1,2,3)}. How many tick marks per
"decade?"}

\item{...}{Additional arguments passed to
\code{\link[ggplot2]{scale_y_continuous}}}
}
\description{
Y-axis wrapper for scale_ratio
}
\examples{
# examples needed here

}
