% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{scale_y_ratio}
\alias{scale_y_ratio}
\title{Ratio-based position scales for continuous data (x & y)}
\usage{
scale_y_ratio(tickVal = "divMult", trans = "log", ...)
}
\arguments{
\item{tickVal}{Character, one of "divMult", "nel", "centiNel", or "propChange"}

\item{trans}{Function or Character name of transformation, most likely "log"}

\item{...}{Additional arguments passed to scale_y_continuous}
}
\description{
\code{scale_x_ratio} and \code{scale_y_ratio} are alternatives to  traditional
\verb{scale_*_continuous} scales for continuous x and y aesthetics, to explicitly
highlight multiplicative or geometric value changes. Rather than traditional
log transformations (as in \verb{scale_*_log10()}), which rescale the axis and
return tickmarks on the original scale of the data, \verb{scale_*_ratio} axis tick
values represent a multiplicative change from a reference point. These scales
may be especially useful for highlighting proportional changes.
}
\examples{
y <- exp(seq(-2,5, length.out = 10))
x <- 1:10
dat <- data.frame(x, y)
dat \%>\% ggplot2::ggplot(ggplot2::aes(x, y))+
     ggplot2::geom_point()+
     ggplot2::geom_hline(yintercept = 1, size = 0.2) +
     scale_y_ratio()
}
