% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioScales.R
\name{split_decades}
\alias{split_decades}
\title{Split stingy limit_breaks into three parts per complete decade}
\usage{
split_decades(v)
}
\arguments{
\item{v}{vector on the unlogged scale to be examined and split}
}
\value{
vector with splits added
}
\description{
Split stingy limit_breaks into three parts per complete decade
}
